feat: add navigation between pages
search and results move to Home page
